+++ tests/unit/test_config.py
@@ -41,13 +41,54 @@ class TestConfigPriorityToDifficulty:
         assert settings.priority_to_difficulty == _DEFAULT_PRIORITY_TO_DIFFICULTY
 
     @staticmethod
-    def should_accept_keys_as_enum_values():
-        settings = Settings(priority_to_difficulty={4: 2, 3: 1.5, 2: 1, 1: 0.1})
+    def should_accept_keys_as_direct_enum_values():
+        settings = Settings(priority_to_difficulty={4: "2", 3: "1.5", 2: "1", 1: "0.1"})
+        assert settings.priority_to_difficulty == _DEFAULT_PRIORITY_TO_DIFFICULTY
+
+    @staticmethod
+    def should_accept_keys_as_indirect_enum_values():
+        settings = Settings(priority_to_difficulty={"4": 2, "3": 1.5, "2": 1, "1": 0.1})
         assert settings.priority_to_difficulty == _DEFAULT_PRIORITY_TO_DIFFICULTY
 
     @staticmethod
     def should_have_default_values():
         assert Settings().priority_to_difficulty == _DEFAULT_PRIORITY_TO_DIFFICULTY
+
+
+class TestConfigLabelToDifficulty:
+    @staticmethod
+    def should_accept_case_insensitive_labels():
+        settings = Settings(
+            label_to_difficulty={
+                "Urgent": "hard",
+                "Important": "medium",
+                "Normal": "easy",
+                "Low": "trivial",
+            }
+        )
+        assert settings.label_to_difficulty == {
+            "urgent": HabiticaDifficulty.HARD,
+            "important": HabiticaDifficulty.MEDIUM,
+            "normal": HabiticaDifficulty.EASY,
+            "low": HabiticaDifficulty.TRIVIAL,
+        }
+
+    @staticmethod
+    def should_allow_missing_difficulty_values():
+        settings = Settings(label_to_difficulty={"Urgent": "Hard"})
+        assert settings.label_to_difficulty == {"urgent": HabiticaDifficulty.HARD}
+
+    @staticmethod
+    def should_allow_multiple_identical_difficulty_values():
+        settings = Settings(label_to_difficulty={"Urgent": "Hard", "Important": "Hard"})
+        assert settings.label_to_difficulty == {
+            "urgent": HabiticaDifficulty.HARD,
+            "important": HabiticaDifficulty.HARD,
+        }
+
+    @staticmethod
+    def should_have_default_empty_label_to_difficulty():
+        assert Settings().label_to_difficulty == {}
 
 
 class TestGetSettings: